requirements.txt
discord.py>2.3
wavelink>2.6
python-dotenv>=1.0
jishaku>=2.5